requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
environs==10.3.0
et-xmlfile==1.1.0
idna==3.6
marshmallow==3.20.2
openpyxl==3.1.2
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.0
